unwinding::abi::custom_eh_frame_finder

Function set_custom_eh_frame_finder

pub fn set_custom_eh_frame_finder(
    fde_finder: &'static (dyn EhFrameFinder + Sync),
) -> Result<(), SetCustomEhFrameFinderError>

Sets the global EH frame finder.

This function should only be called once during the lifetime of the program.

Errors

An error is returned if this function has already been called during the lifetime of the program.